Identify first argument of pulldata by splitting on comma
The first argument has to be static for any of this to work so we can assume that it doesn't contain commas. That is, we're using the value of the first argument as the instance id so it has to be a string and can't be something like a concat expression.